chore: Update go.mod dependencies
This commit updates the go.mod file to include the latest version of the fyne.io/systray package. It also removes the github.com/shirou/gopsutil and github.com/go-ole/go-ole dependencies, which are no longer needed. This change ensures that the project has the most up-to-date dependencies and removes unnecessary dependencies.